solver options: push _effective_solver_options to SPOpt; migrate xhat_eval + lshaped
Completes the migration of internal readers off current_solver_options.

## SPOpt now owns the layer state

solver_options_layers, _dynamic_solver_options_layer,
current_solver_options, and the _effective_solver_options(k)
method are now all set up in SPOpt.__init__, with the same
legacy-iter0/iterk-dict synthesis shim (and rank-gated
DeprecationWarning) that previously lived in PHBase.__init__.
PHBase inherits everything from SPOpt; its init no longer
duplicates the setup. The deprecated iter0_solver_options /
iterk_solver_options @property shims stay on PHBase since they
are PH-flavored (the iter0/iterk predicate names) and not
meaningful for non-PH SPOpt subclasses like Xhat_Eval.

## Xhat_Eval and lshaped_bounder migrated

* mpisppy/utils/xhat_eval.py:373 — Xhat_Eval is a steady-state
  evaluator with no per-PH-iter distinction of its own; its
  internal solve_loop now uses
  self._effective_solver_options(1), folding any iterk /
  starting_at_iter layers the user configured. The fold's final
  layer is still current_solver_options, preserving the
  back-compat overlay path for any external mutator.
* mpisppy/cylinders/lshaped_bounder.py:30 — iter0 prep solve
  now uses self.opt._effective_solver_options(0) for the iter0
  fold.
* mpisppy/cylinders/lshaped_bounder.py:53 — the reassignment
  `self.opt.current_solver_options = self.opt.options
  ["iterk_solver_options"]` is now unnecessary and removed;
  Xhat_Eval pulls the iterk fold directly each solve.

## Status

With #702 (in flight) plus this PR, every internal reader of
current_solver_options is now off it. The attribute remains
only as the back-compat overlay read by _effective_solver_options
itself. Read-side deprecation is the natural next step.

§6.3 in the design doc updated to reflect the new state and the
location move (PHBase → SPOpt).

Local: 237/238 pass across the wider sweep (the one failure is
the pre-existing unrelated test_proper_bundler flake on main).
ruff clean.
