+++ doc/designs/solver_options_redesign.md
@@ -841,7 +841,7 @@ Action items (all complete):
 |----------------------------------------------------------|-------------------------------------|----------------------------------------------------------------------------------|---------------|
 | `options["iter0_solver_options"]`, `options["iterk_solver_options"]` (input to PHBase) | **DeprecationWarning fires when non-empty; legacy dicts still folded into `solver_options_layers`** | unchanged target (`solver_options_layers`) | future, TBD   |
 | `PHBase.iter0_solver_options`, `iterk_solver_options` (attribute reads) | **DeprecationWarning fires on every read; both are property shims returning `fold_solver_options_layers(layers, 0|1)`** | use `_effective_solver_options(k)` | future, TBD |
-| `PHBase.current_solver_options` (attribute reads / writes) | active; still used internally as the back-compat dynamic-overrides overlay by several spokes (fwph, aph, xhat_eval, cross_scen, lshaped) | not deprecated in phase 8 — pending those spokes migrating to the layer system | future, TBD |
+| `SPOpt.current_solver_options` (attribute reads / writes) | all internal readers migrated: fwph, aph, cross_scen_extension (PHBase-based) and xhat_eval, lshaped_bounder (SPOpt-based) now read via `_effective_solver_options(k)`; the attribute remains as the back-compat dynamic-overrides overlay path read by `_effective_solver_options` itself | reads can be deprecated next (only the back-compat overlay reads it now); writes (Gapper auto-mode, spoke iter0→iterk handoffs) can be migrated to the dynamic_gapper layer | future, TBD |
 | `option_string_to_dict`                                  | active                              | unchanged                                                                         | n/a           |
 | `option_dict_to_string`                                  | active                              | unchanged                                                                         | n/a           |
